+++ src/xdcheckin/core/chaoxing.py
@@ -23,6 +23,7 @@ class Chaoxing:
 		"requests_cache_enabled": True,
 		"chaoxing_course_get_activities_courses_limit": 48,
 		"chaoxing_checkin_location_address_override": False,
+		"chaoxing_checkin_location_address_override_maxlen": 0,
 		"chaoxing_checkin_location_randomness": True
 	}
 
@@ -484,7 +485,7 @@ class Chaoxing:
 				"latitude": _randomness(location_new["latitude"]),
 				"longitude": _randomness(location_new["longitude"])
 			})
-		if self.config["chaoxing_checkin_location_address_override"] and location["address"]:
+		if self.config["chaoxing_checkin_location_address_override"] and len(location["address"]) <= self.config["chaoxing_checkin_location_address_override_maxlen"] or len(location["address"]):
 			location_new["address"] = location["address"]
 		return location_new