+++ src/xdcheckin/server/static/location.js
@@ -13,14 +13,11 @@ async function setLocation(loc, name) {
 async function inputLocation() {
 	let latitude, longitude, address;
 	if (!(latitude = prompt("Input latitude:")) || isNaN(latitude))
-		return ((latitude === null) ? undefined :
-					      alert("Invalid input."));
+		return latitude === null ? undefined : alert("Invalid input.");
 	if (!(longitude = prompt("Input latitude:")) || isNaN(longitude))
-		return ((longitude === null) ? undefined :
-					       alert("Invalid input."));
+		return longitude === null ? undefined : alert("Invalid input.");
 	if (!(address = prompt("Input address:")) || /[()\[\]{}]/.test(address))
-		return ((address === null) ? undefined :
-					     alert("Invalid input."));
+		return address === null ? undefined : alert("Invalid input.");
 	setLocation({
 		"latitude": latitude,
 		"longitude": longitude,
@@ -29,23 +26,16 @@ async function inputLocation() {
 }
 
 async function listLocations() {
-	if (listLocations.calling)
-		return;
-	listLocations.calling = true;
 	let e = document.getElementById("locations-list-div");
 	e.appendChild(newElement("button", {
 		innerText: "Input",
-		onclick: () => {
-			inputLocation();
-		}
+		onclick: inputLocation
 	}));
 	for (let buildingName in g_locations) {
 		e.appendChild(newElement("button", {
 			innerText: buildingName,
-			onclick: () => {
-				setLocation(g_locations[buildingName],
-					    buildingName);
-			},
+			onclick: () => setLocation(g_locations[buildingName],
+						   buildingName),
 		}));
 	}
 	let loc = localStorage.getItem('location_');
@@ -55,5 +45,4 @@ async function listLocations() {
 	else
 		setLocation(JSON.parse(loc),
 			    localStorage.getItem('location_name'));
-	listLocations.calling = false;
 }